# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News project API",
        "version": "1"
    },
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get edit profile view",
                "description": "Retrieve the edit profile view with data required for the form.",
                "operationId": "8cfcc99d8d1aa366314cd8abf9cd9c9a",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "allLanguageNames": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string",
                                                "property": null
                                            }
                                        },
                                        "allCountryNames": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string",
                                                "property": null
                                            }
                                        },
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update the user's profile information",
                "description": "Updates the authenticated user's profile information.",
                "operationId": "updateUserProfile",
                "requestBody": {
                    "description": "User's profile information to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "language": {
                                        "type": "string"
                                    },
                                    "country": {
                                        "type": "string"
                                    },
                                    "categories": {
                                        "type": "string"
                                    },
                                    "current_password": {
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirects to the profile edit page with a success message.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "profile-updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action, user is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors in the request."
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Delete the user's account.",
                "description": "Delete the user's account. Requires current user's password.",
                "operationId": "1d6a5c6489e35a763aa97ebef12863ce",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "description": "The current user's password.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User account has been deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User account has been deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized access"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "password": [
                                                    "The current password is invalid."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server error occurred while deleting the user's account."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "name": {
                        "description": "User name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string"
                    },
                    "country": {
                        "description": "User country",
                        "type": "string"
                    },
                    "language": {
                        "description": "User language",
                        "type": "string"
                    },
                    "category": {
                        "description": "User category",
                        "type": "string"
                    },
                    "role": {
                        "description": "User role",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Date and time of user creation",
                        "type": "timestamp",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Date and time of user update",
                        "type": "timestamp",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "favourites": {
                        "description": "User's favourite news articles",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserFavourite"
                        }
                    },
                    "comments": {
                        "description": "User's comments on news articles",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserComment"
                        }
                    },
                    "logs": {
                        "description": "User's logs of activity on the news portal",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserLog"
                        }
                    }
                },
                "type": "object"
            },
            "UserComment": {
                "properties": {
                    "uuid": {
                        "type": "string",
                        "readOnly": true,
                        "example": "f6ebff50-aa5d-11eb-aebf-0242ac130002"
                    },
                    "userUuid": {
                        "type": "string",
                        "example": "f6ebff50-aa5d-11eb-aebf-0242ac130002"
                    },
                    "url": {
                        "type": "text",
                        "example": "https://example.com/article"
                    },
                    "commentText": {
                        "type": "text",
                        "example": "This is a comment"
                    },
                    "created_at": {
                        "type": "timestamp",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2023-05-08T12:30:00Z"
                    },
                    "updated_at": {
                        "type": "timestamp",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2023-05-08T12:30:00Z"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "UserFavourite": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": "true"
                    },
                    "title": {
                        "type": "string"
                    },
                    "url": {
                        "type": "text"
                    },
                    "author": {
                        "type": "string"
                    },
                    "description": {
                        "type": "text"
                    },
                    "imageUrl": {
                        "type": "text"
                    },
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "created_at": {
                        "type": "timestamp",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "timestamp",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "UserLog": {
                "title": "UserLog",
                "description": "User log model",
                "properties": {
                    "id": {
                        "description": "The unique uuid for the log entry",
                        "type": "string",
                        "format": "uuid"
                    },
                    "action": {
                        "description": "The action performed by the user",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the action performed by the user",
                        "type": "text"
                    },
                    "userId": {
                        "description": "The UUID of the user who performed the action",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The date and time when the log entry was created",
                        "type": "timestamp",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time when the log entry was last updated",
                        "type": "timestamp",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}